+++ app/actions/aspaceSearch.ts
@@ -7,6 +7,7 @@ import {
 } from '@kenxirwin/archives-space-api-client';
 import logger from '@/lib/logger';
 import callNumberOverrides from '@/lib/catalogs/aspace/callNumberOverrides';
+import titleOverrides from '@/lib/catalogs/aspace/titleOverrides';
 import type {
   RepoArchivalObject,
   RepoResources,
@@ -18,8 +19,6 @@ import type {
   CatalogSearchResult,
 } from '@/lib/catalogs/types';
 import { ZodError } from 'zod';
-import { findNodeByKeyValuePair } from '@/lib/findNodeByKeyValuePair';
-import { cli } from 'winston/lib/winston/config';
 import { HandleMissingInstances } from './aspaceHandleMissingInstances';
 
 export interface ArchivalObjectExtraInfo {
@@ -82,6 +81,7 @@ export async function searchByUrl(url: string, client: AspaceClient) {
     const raw = await client.getUrl(url, {
       resolve: ['linked_agents', 'repository', 'top_container'],
     });
+    logger.silly(`Raw response: ${JSON.stringify(raw)}`);
     let extraInfo = {};
     switch (true) {
       // https://archivesstaff.lib.miamioh.edu/api/repositories/2/resources/634
@@ -120,9 +120,9 @@ export async function searchByUrl(url: string, client: AspaceClient) {
         */
         if (parsed.instances.length == 0) {
           const parentResourceUrl = parsed.resource.ref;
-          console.log(`parentResourceUrl: ${parentResourceUrl}`);
+          logger.verbose(`parentResourceUrl: ${parentResourceUrl}`);
           const originalUri = url.match(/\/repositories\/.*/);
-          console.log(`originalUri: ${originalUri}`);
+          logger.verbose(`originalUri: ${originalUri}`);
           if (originalUri !== null) {
             extraInfo = await HandleMissingInstances(
               originalUri.toString(),
@@ -131,33 +131,6 @@ export async function searchByUrl(url: string, client: AspaceClient) {
             );
           }
         }
-        //   console.log(`looking for more info about ${originalUri}`);
-        //   const resourceWithTree: RepoResources = await getResourceTree(
-        //     parentResourceUrl,
-        //     client,
-        //   );
-        //   console.log(`Resource Title: ${resourceWithTree.title}`);
-        //   console.log(
-        //     `find key value pair in tree with record_uri: "${originalUri?.toString()}"`,
-        //   );
-        //   const node = findNodeByKeyValuePair(
-        //     resourceWithTree,
-        //     'record_uri',
-        //     originalUri?.toString(), // not sure why toString is needed, but it is
-        //   );
-        //   const numItems = node.children.length;
-        //   const firstItemUrl = apiBaseUrl + node.children[0].record_uri;
-        //   const firstRecordArgusData = await searchByUrl(firstItemUrl, client);
-        //   console.log(`numItems: ${numItems}`);
-        //   console.log(`first child = ${JSON.stringify(node.children[0])}`);
-        //   extraInfo = {
-        //     numItems,
-        //     firstItemUrl,
-        //     firstRecordArgusData,
-        //   };
-        // }
-        // console.log(`extraInfo: ${JSON.stringify(extraInfo)}`);
-        /* End special section dealing with missing archival_object data */
 
         let argusData;
         if (Object.keys(extraInfo).length > 0) {
@@ -202,7 +175,7 @@ function repoResourcesToDraft(data: RepoResources, url: string) {
     callNumber:
       callNumberOverrides.resources?.bib?.(data) ??
       `${data.id_0}--${data.id_1}--${data.id_2}`,
-    itemTitle: data.title,
+    itemTitle: titleOverrides.resources?.(data) ?? data.title,
     catalog: 'ASPACE',
     catalogId: data.uri,
     catalogIdType: 'uri',
@@ -242,7 +215,7 @@ function repoTopContainerToDraft(data: RepoTopContainer, url: string) {
   const bibData: BibDataDraft = {
     author: 'Unknown',
     callNumber: callNumberOverrides.topContainer?.bib?.(data) ?? data.indicator,
-    itemTitle: data.long_display_string,
+    itemTitle: titleOverrides.topContainer?.(data) ?? data.long_display_string,
     catalog: 'ASPACE',
     catalogId: data.uri,
     catalogIdType: 'uri',
@@ -286,11 +259,11 @@ const getItems = (data: RepoArchivalObject) => {
       box: '',
       call_number:
         callNumberOverrides.archivalObject?.item?.(item, data) ??
-        item.sub_container.top_container._resolved?.display_string ??
+        item.sub_container?.top_container?._resolved?.display_string ??
         '',
       copy_id: '',
       description:
-        item.sub_container.top_container._resolved?.long_display_string ?? '',
+        item.sub_container?.top_container?._resolved?.long_display_string ?? '',
       folder: '',
       location_code: data.repository._resolved?.slug ?? '',
       location_name: data.repository._resolved?.name ?? '',
@@ -335,7 +308,7 @@ function repoArchivalObjectToDraft(
         )
         .join('; ') ??
       '',
-    itemTitle: data.title,
+    itemTitle: titleOverrides.archivalObject?.(data) ?? data.title,
     catalog: 'ASPACE',
     catalogId: data.uri,
     catalogIdType: 'uri',